Reorganise proj structure and update imports
